Interacting with SCC behind proxy and SMT
When the user try to get the products from the SCC behind a proxy or
from a SMT server, container-suseconnect fails. This happens because it
checks if it need to get the regcodes just comparing the URL. Then, if
the user has a proxy, the URL will not be the expected one, skipping the
code to get the regcodes and by consequence, it cannot get the products
list.

Avoiding add more configuration option, this commit fixes this issue
always requesting the regcodes.  If the user uses a SMT server, which
does not have this API. The container-suseconnect assume it is a SMT
server and continue without the regcodes.

Issue #28
bsc#1127030

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>