susebuild: Add SCC credentials to returned info
Add these credentials to the URLs returned by the urlresolve plugin.
The ones that are mapped into the container might not be up to date when
running inside a public cloud instance. This also makes the plugin
drop the "credentials" parameter as that is not needed anymore with
the userinfo present in the URL.