Rework ReadConfigFromServer() to use json.Decoder
Using Read() directly has the risk of doing a short read. E.g. on a busy
machine it might not read an entire JSON Object in one call. Switch the
implementation to use json.Decoder and rely on it doing the Right Thing (TM)

Adjust the Test Case accordingly to check for the correct error message.