Add flag --log-credentials-errors
Currently container-suseconnect will print errors to the console when it fails
to obtain the SLE credentials. This is however causing UX issues in the SLE BCI
containers, which work perfectly fine without a SLE subscription.

Thus a new flag is introduced: `--log-credentials-errors`. When this flag is set
or if the environment variable `CONTAINER_SUSECONNECT_LOG_CREDENTIALS_ERR` is
set, then `container-suseconnect` will continue printing errors as it did, but
the default is now that errors with obtaining the credentials will be silently
swallowed so that users of SLE BCI don't get confused by this error message.

This fixes https://github.com/SUSE/container-suseconnect/issues/62